feat: Add support for Google, Facebook, and Line OAuth2 providers
This commit adds support for Google, Facebook, and Line OAuth2 providers to the project. It includes the necessary dependencies and configurations for each provider, as well as a new `Provider` struct to handle the authentication process. The `Provider` struct has methods to generate an authorization URL and exchange the authorization code for an access token. Overall, this commit lays the foundation for additional OAuth2 providers to be added in the future.